requirements.txt
# Requires Python 3.9-3.12 (PaddlePaddle constraint)
pandas
numpy
requests
pdf2image
psutil
paddlepaddle>=3.0,<4.0
paddleocr>=3.0,<4.0
rank-bm25
qdrant-client
transformers
torch
Pillow